requirements.txt
streamlit==1.28.1
mysql-connector-python==8.1.0
pandas==2.1.1
python-dotenv==1.0.0
SpeechRecognition==3.10.0
langchain==0.0.330
langchain-google-genai==0.0.6
langchain-groq==0.0.1

